requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-WTF==1.2.1
email-validator==2.1.0.post1
bcrypt==4.1.2
pyotp==2.9.0
python-dotenv==1.0.1
Pillow==10.2.0
captcha==0.5.0 